- fix: ImageSharpenFS and all frequency separation nodes output float64 tensors causing dtype mismatch downstream (#219, #229)   - Cast outputs to float32 in `Frequency_Separation_Hard_Light`, `Frequency_Separation_Hard_Light_LAB`,     `Frequency_Separation_Hard_Light_Vivid`, `Frequency_Separation_Linear_Light`, `Frequency_Separation_FFT`,     and `RGB_to_LAB` classes in `images.py`   - Internal processing stays float64 for numerical precision; only outputs are cast   - Also fixes MPS (Mac) compatibility since MPS doesn't support float64
- fix: scheduler list concatenation breaks references held by SwarmUI and other tools (#220)
  - Changed `list = list + ["beta57"]` to `list.append("beta57")` in 4 locations in `res4lyf.py`
  - Preserves in-place references so schedulers registered after RES4LYF are still visible

- fix: FlashAttention custom_op registration causes dispatcher assert on Torch 2.9+ (#225, #210)
  - Added check for already-registered `torch.ops.flash_attention.flash_attn` before calling `@torch.library.custom_op`
  - If already registered, uses plain wrapper function instead of re-registering
  - Broadened exception catch from `AttributeError` to `Exception` for schema assertion errors
  - File: `sd/attention.py`

- fix: SyntaxWarning from invalid escape sequence `\h` in Python 3.12+ (#185)
  - Changed `"$Δ \hat{t}$"` to `r"$Δ \hat{t}$"` (raw string) in `helper_sigma_preview_image_preproc.py`

+++ images.py
@@ -474,7 +474,7 @@ class Frequency_Separation_Hard_Light:
         if original is None:
             original = hard_light_blend(low_pass.to(torch.float64).to('cuda'), high_pass.to(torch.float64).to('cuda'))
 
-        return (high_pass, original, low_pass,)
+        return (high_pass.to(torch.float32), original.to(torch.float32), low_pass.to(torch.float32),)
 
 
 class Frequency_Separation_Hard_Light_LAB:
@@ -524,7 +524,7 @@ class Frequency_Separation_Hard_Light_LAB:
             lab_original = denormalize_lab(lab_original_normalized)
             original = lab_to_rgb(lab_original).permute(0, 2, 3, 1)
 
-        return (high_pass, original, low_pass)
+        return (high_pass.to(torch.float32), original.to(torch.float32), low_pass.to(torch.float32))
     
     
 class Frame_Select:
@@ -639,7 +639,7 @@ class Image_Channels_LAB:
             LAB = torch.cat([L,A,B], dim=1)
             RGB = lab_to_rgb(LAB.to(torch.float64)).permute(0,2,3,1)
 
-        return (RGB, L, A, B,)
+        return (RGB.to(torch.float32), L.to(torch.float32), A.to(torch.float32), B.to(torch.float32),)
     
     
 
@@ -671,7 +671,7 @@ class Frequency_Separation_Vivid_Light:
         if original is None:
             original = hard_light_blend(high_pass.to(torch.float64), low_pass.to(torch.float64))
 
-        return (high_pass, original, low_pass,)
+        return (high_pass.to(torch.float32), original.to(torch.float32), low_pass.to(torch.float32),)
 
 
 class Frequency_Separation_Linear_Light:
@@ -703,7 +703,7 @@ class Frequency_Separation_Linear_Light:
         if original is None:
             original = linear_light_blend(low_pass.to(torch.float64).to('cuda'), high_pass.to(torch.float64).to('cuda'))
 
-        return (high_pass, original, low_pass,)
+        return (high_pass.to(torch.float32), original.to(torch.float32), low_pass.to(torch.float32),)
 
 
 class Frequency_Separation_FFT:
@@ -737,7 +737,7 @@ class Frequency_Separation_FFT:
         if original is None:
             original = low_pass + high_pass
 
-        return (high_pass, original, low_pass,)
+        return (high_pass.to(torch.float32), original.to(torch.float32), low_pass.to(torch.float32),)
